requirements.txt
fastapi==0.110.0
pydantic==2.5.3
uvicorn==0.24.0
pymongo==4.6.0
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.9
pandas==2.1.4
openpyxl==3.1.2
reportlab==4.0.7
cloudinary==1.39.0
httpx==0.24.1
pytest==7.4.3
email-validator==2.1.0
Jinja2==3.1.3